pytorch_imple/requirements.txt
albumentations==1.4.22
matplotlib==3.9.3
natsort==8.4.0
numpy==2.2.0
opencv_python_headless==4.10.0.84
Pillow==11.0.0
pycocotools==2.0.8
Requests==2.32.3
torch==2.3.1+cu121
torch_lr_finder==0.2.2
torchinfo==1.8.0
torchvision==0.18.1+cu121
tqdm==4.67.1
